Extract properties accordion into a new metacorp modal component
